Fin. (see full commit message for final details)
Determined that the last "out of range" exception was from bad input
(again). This time, it was a term that had a hyphen instead of a dash.
So it goes.

The output file has been included for before and after. I've imported it
into my Anki flashcard app (the original purpose for this exercise) and
it worked like a charm! I can now get ready for the 2017 season and
beyond.